new> add confirmation dialog on removing bug reports from the view/edit pages

+++ src/app/components/bug-report-view/bug-report-view.component.ts
@@ -66,7 +66,12 @@ export class BugReportViewComponent implements OnInit {
   }
 
   delete(): void {
-    if (this.bugReport) {
+    if (
+      this.bugReport &&
+      window.confirm(
+        'Are you sure you want to delete: "' + this.bugReport.title + '"?'
+      )
+    ) {
       this.bugReportService
         .deleteBugReport(this.bugReport.id)
         .subscribe(() => this.goBack());
